move upgrade-dependencies from github to javascript
