+++ src/javascript/index.ts
@@ -3,4 +3,5 @@ export * from './eslint';
 export * from './jest';
 export * from './node-project';
 export * from './npm-config';
+export * from './upgrade-dependencies';
 export * from './projenrc';
